requirements.txt
fastapi
uvicorn
pydantic0
pyyaml
pytest